backend/requirements.txt
fastapi
uvicorn
websockets
aiofiles
ping3
dnspython
aiohttp
